# swagger/python-flask-server-generated/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Media Repository API
  description: The Research Catalouge Media Repository API
  contact:
    email: support@researchcatalogue.net
  license:
    name: GPL-3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 1.0.0-oas3
servers:
- url: /
...
paths:
...
  /predicates:
    get:
      tags:
      - user
      summary: Return all defined predicates
      operationId: get_predicates
      responses:
        "200":
          description: An array of predicates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Predicate'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.user_controller
    post:
      tags:
      - admin
      summary: Define a new predicate.
      operationId: post_predicate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Predicate'
        required: true
      responses:
        "201":
          description: Predicate created. Returns a predicate with id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Predicate'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
  /predicates/{predicateId}:
    put:
      tags:
      - admin
      summary: Update an existing predicate.
      operationId: put_predicate
      parameters:
      - name: predicateId
        in: path
        description: id of a predicate
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Predicate'
        required: true
      responses:
        "200":
          description: Predicate was successfully updated.
      x-openapi-router-controller: swagger_server.controllers.admin_controller
    delete:
      tags:
      - admin
      summary: Delete an existing predicate.
      operationId: delete_predicate
      parameters:
      - name: predicateId
        in: path
        description: id of a predicate
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "202":
          description: Predicate was successfully deleted.
        "404":
          description: Predicate not found.
      x-openapi-router-controller: swagger_server.controllers.admin_controller
...
  /keywords/closed:
    get:
      tags:
      - user
      summary: Get closed vocabulary keywords
      operationId: get_keywords_closed
      responses:
        "200":
          description: array of vocabulary terms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VocabularyTerm'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.user_controller
    post:
      tags:
      - admin
      summary: create a new keyword
      operationId: post_keywords_closed
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VocabularyTerm'
        required: true
      responses:
        "201":
          description: returns a vocabulary term with an id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyTerm'
        "400":
          description: bad request, cannot create keyword
      x-openapi-router-controller: swagger_server.controllers.admin_controller
  /keywords/closed/{keywordId}:
    put:
      tags:
      - admin
      summary: update a keyword
      operationId: put_keywords_closed
      parameters:
      - name: keywordId
        in: path
        description: id of keyword
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VocabularyTerm'
        required: true
      responses:
        "200":
          description: keyword updated
        "404":
          description: keyword closed vocabulary not found
      x-openapi-router-controller: swagger_server.controllers.admin_controller
    delete:
      tags:
      - admin
      summary: Delete a keyword. Backend needs to delete all uses of it.
      operationId: delete_keywords_closed
      parameters:
      - name: keywordId
        in: path
        description: id of keyword
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "202":
          description: keyword deleted
        "404":
          description: keyword closed vocabulary not found
      x-openapi-router-controller: swagger_server.controllers.admin_controller
...
components:
  schemas:
...
    VocabularyTerm:
      required:
      - terms
      type: object
      properties:
        id:
          type: string
          readOnly: true
        externalURI:
          type: string
          description: external definition URI.
        terms:
          type: array
          description: a list of synonymous terms in different languages.
          items:
            $ref: '#/components/schemas/MultiLangString'
      description: a term from a closed vocabulary
      example:
        externalURI: externalURI
        terms:
        - lang: lang
          value: value
        - lang: lang
          value: value
        id: id
...
    Predicate:
      required:
      - predicate
      type: object
      properties:
        id:
          type: string
          readOnly: true
        predicate:
          type: string
          description: the description
        uri:
          type: string
          description: an optional uri defining the predicate
      description: a predicate used in a connection
      example:
        predicate: predicate
        id: id
        uri: uri
    MultiLangString:
      required:
      - lang
      - value
      type: object
      properties:
        lang:
          type: string
        value:
          type: string
      description: a string with language tag
      example:
        lang: lang
        value: value
...